# apps/itsm/classic/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ITSM Classic",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/incidents/{iid}/link/change": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Link Change",
        "operationId": "linkIncidentChange",
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/changes": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Create Change",
        "operationId": "createChange",
        "parameters": [
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Change"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "changes"
        ],
        "summary": "List Changes",
        "operationId": "listChanges",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Change"
                  },
                  "title": "Response Listchanges"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/changes/{cid}": {
      "get": {
        "tags": [
          "changes"
        ],
        "summary": "Get Change",
        "operationId": "getChange",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Change"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/changes/{cid}/submit": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Submit Change",
        "operationId": "submitChange",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Change"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/changes/{cid}/approve": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Approve Change",
        "operationId": "approveChange",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Change"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/changes/{cid}/reject": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Reject Change",
        "operationId": "rejectChange",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Change"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/changes/{cid}/schedule": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Schedule Change",
        "operationId": "scheduleChange",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Change"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/changes/{cid}/implement": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Implement Change",
        "operationId": "implementChange",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Change"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/changes/{cid}/complete": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Complete Change",
        "operationId": "completeChange",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Change"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/changes/{cid}/rollback": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Rollback Change",
        "operationId": "rollbackChange",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Change"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/changes/{cid}/close": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Close Change",
        "operationId": "closeChange",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Change"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/changes/{cid}/link/ci": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Link Change Ci",
        "operationId": "linkChangeCi",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCiRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Change"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/changes/{cid}/link/incident": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Link Change Incident",
        "operationId": "linkChangeIncident",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkIncidentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Change"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Change": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "type": {
            "type": "string",
            "enum": [
              "normal",
              "emergency"
            ],
            "title": "Type"
          },
          "risk": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Risk"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "submitted",
              "cab_approved",
              "scheduled",
              "implementing",
              "implemented",
              "closed",
              "rejected",
              "rolled_back"
            ],
            "title": "Status"
          },
          "requester": {
            "type": "string",
            "title": "Requester"
          },
          "implementer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implementer"
          },
          "affected_cis": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Cis"
          },
          "linked_incidents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Linked Incidents"
          },
          "change_window": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Window"
          },
          "rollback_plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rollback Plan"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/HistoryEntry"
            },
            "type": "array",
            "title": "History"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Actions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "type",
          "risk",
          "status",
          "requester",
          "created_at",
          "updated_at"
        ],
        "title": "Change"
      },
      "Comment": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "author",
          "body",
          "at"
        ],
        "title": "Comment"
      },
      "CreateChangeRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "type": {
            "type": "string",
            "enum": [
              "normal",
              "emergency"
            ],
            "title": "Type",
            "default": "normal"
          },
          "risk": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Risk",
            "default": "medium"
          },
          "affected_cis": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Cis"
          },
          "rollback_plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rollback Plan"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "CreateChangeRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryEntry": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "from_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Status"
          },
          "to_status": {
            "type": "string",
            "title": "To Status"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          }
        },
        "type": "object",
        "required": [
          "action",
          "to_status",
          "at",
          "actor"
        ],
        "title": "HistoryEntry"
      },
      "Incident": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "priority": {
            "type": "string",
            "enum": [
              "P1",
              "P2",
              "P3",
              "P4"
            ],
            "title": "Priority"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "in_progress",
              "pending_user",
              "resolved",
              "closed",
              "escalated"
            ],
            "title": "Status"
          },
          "category": {
            "type": "string",
            "enum": [
              "hardware",
              "software",
              "network",
              "security",
              "performance"
            ],
            "title": "Category"
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "requester": {
            "type": "string",
            "title": "Requester"
          },
          "affected_ci": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affected Ci"
          },
          "linked_cis": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Linked Cis"
          },
          "linked_changes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Linked Changes"
          },
          "comments": {
            "items": {
              "$ref": "#/components/schemas/Comment"
            },
            "type": "array",
            "title": "Comments"
          },
          "resolution_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "sla_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/HistoryEntry"
            },
            "type": "array",
            "title": "History"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Actions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "priority",
          "status",
          "category",
          "requester",
          "created_at",
          "updated_at"
        ],
        "title": "Incident"
      },
      "LinkChangeRequest": {
        "properties": {
          "change_id": {
            "type": "string",
            "title": "Change Id"
          }
        },
        "type": "object",
        "required": [
          "change_id"
        ],
        "title": "LinkChangeRequest"
      },
      "LinkCiRequest": {
        "properties": {
          "ci_id": {
            "type": "string",
            "title": "Ci Id"
          }
        },
        "type": "object",
        "required": [
          "ci_id"
        ],
        "title": "LinkCiRequest"
      },
      "LinkIncidentRequest": {
        "properties": {
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          }
        },
        "type": "object",
        "required": [
          "incident_id"
        ],
        "title": "LinkIncidentRequest"
      },
...
      "ScheduleChangeRequest": {
        "properties": {
          "change_window": {
            "type": "string",
            "title": "Change Window"
          }
        },
        "type": "object",
        "required": [
          "change_window"
        ],
        "title": "ScheduleChangeRequest"
      },
...
    }
  }
}